# feat(perf-issues): Switch span hashing strategy to use parametrized DB queries