--- a/src/sentry/event_manager.py
+++ b/src/sentry/event_manager.py
@@ -120,7 +120,6 @@ from sentry.ratelimits.sliding_windows import Quota, RedisSlidingWindowRateLimit
 from sentry.reprocessing2 import is_reprocessed_event, save_unprocessed_event
 from sentry.shared_integrations.exceptions import ApiError
 from sentry.signals import first_event_received, first_transaction_received, issue_unresolved
-from sentry.spans.grouping.strategy.config import INCOMING_DEFAULT_CONFIG_ID
 from sentry.tasks.commits import fetch_commits
 from sentry.tasks.integrations import kick_off_status_syncs
 from sentry.tasks.process_buffer import buffer_incr
@@ -2166,23 +2165,6 @@ def _calculate_span_grouping(jobs: Sequence[Job], projects: ProjectsMapping) -> 
                 "save_event.transaction.span_group_count.default",
                 amount=len(unique_default_hashes),
                 tags={"platform": job["platform"] or "unknown"},
-            )
-
-            # Try the new hashing config that more aggresively parametrizes DB
-            # spans, and record the difference with the default hashing config.
-            with metrics.timer("event_manager.save.get_span_groupings.incoming"):
-                experimental_groupings = event.get_span_groupings(
-                    {"id": INCOMING_DEFAULT_CONFIG_ID}
-                )
-
-            unique_incoming_hashes = set(experimental_groupings.results.values())
-            metrics.incr(
-                "save_event.transaction.span_group_count.incoming",
-                amount=len(unique_incoming_hashes),
-            )
-            metrics.incr(
-                "save_event.transaction.span_group_count.difference",
-                amount=len(unique_default_hashes ^ unique_incoming_hashes),
             )
         except Exception:
             sentry_sdk.capture_exception()