--- a/src/sentry/spans/grouping/strategy/config.py
+++ b/src/sentry/spans/grouping/strategy/config.py
@@ -41,8 +41,7 @@ def register_configuration(config_id: str, strategies: Sequence[CallableStrategy
     CONFIGURATIONS[config_id] = SpanGroupingConfig(config_id, strategy)
 
 
-DEFAULT_CONFIG_ID = "default:2022-10-04"
-INCOMING_DEFAULT_CONFIG_ID = "default:2022-10-27"
+DEFAULT_CONFIG_ID = "default:2022-10-27"
 
 register_configuration(
     "default:2021-08-25",