--- a/tests/sentry/api/serializers/test_event.py
+++ b/tests/sentry/api/serializers/test_event.py
@@ -401,7 +401,7 @@ class DetailedEventSerializerTest(TestCase):
             "causeSpanIds": ["9179e43ae844b174"],
             "desc": "SELECT `books_author`.`id`, `books_author`.`name` FROM "
             "`books_author` WHERE `books_author`.`id` = %s LIMIT 21",
-            "fingerprint": "19e15e0444e0bc1d5159fb07cd4bd2eb",
+            "fingerprint": "e714d718cb4e7d3ce1ad800f7f33d223",
             "offenderSpanIds": [
                 "b8be6138369491dd",
                 "b2d4826e7b618f1b",