Change the default span hashing configuration to `2022-10-27`, which parametrizes SQL queries in DB span descriptions before hashing them. This will start affecting both Suspect Spans and N+1 Performance Issues - Suspect Spans will be better grouped, and the issues with logically-identical but technically-distinct DB queries cauing dupe N+1s should become much rarer.

Changing the hashing strategy affected the span hashes in some tests, so update those too.

Note that the check to ensure DB queries are at least partially parameterized is still in place for now - we'll confirm that this actually improves grouping before removing that restriction.

Also remove the extra metrics added during the transition period to validate the change before switching the strategy over.