@mjq-sentry 2022-11-03
@gggritso:
> For posterity, would be nice to post some stats about the changes in hash counts, and that you verified that this solves specific cases that we've seen (like the Django project one, right?)

For sure! I'm waiting until tomorrow to post the performance and hash count changes to make sure they're representative. I confirmed that the new strategy solves the partially-parametrized query issue earlier by manually applying it but I'll double-check it against this code (should work unmodified now!)

Thanks for the review!