@mjq-sentry 2022-11-07
Comparing the hash calculations since the two grouping strategies started being run:
* 75th percentile: no change
* 95th percentile: time increased ~50% (0.6ms to 0.9ms)
* 99th percentile: time increased ~133% (3ms to 7ms)

So, the long tail is now slower. This makes sense and likely follows the distribution of database spans in transactions - few transactions have many large db query spans, which will be the slowest to run the parametrization code on.

The only material change is to the p99. Given the improvements to suspect spans and N+1 detection (and any future performance issue detectors using db spans) I think this is a reasonable cost. (Note that we've been paying the cost of _both_ the old and new strategies since Thursday and everything is fine - merging this PR will actually decrease the ingest runtime by no longer running the old strategy's calculations).